+++ lib/src/ast/value/string/ml_basic.dart
@@ -2,7 +2,6 @@ library toml.src.ast.value.string.ml_basic;
 
 import 'package:meta/meta.dart';
 import 'package:petitparser/petitparser.dart';
-import 'package:quiver/core.dart';
 
 import '../../../decoder/parser/ranges.dart';
 import '../../../decoder/parser/whitespace.dart';
@@ -182,5 +181,5 @@ class TomlMultilineBasicString extends TomlMultilineString {
       other is TomlMultilineBasicString && value == other.value;
 
   @override
-  int get hashCode => hash3(type, stringType, value);
+  int get hashCode => Object.hash(type, stringType, value);
 }