+++ lib/src/ast/value/string/ml_literal.dart
@@ -2,7 +2,6 @@ library toml.src.ast.value.string.ml_literal;
 
 import 'package:meta/meta.dart';
 import 'package:petitparser/petitparser.dart';
-import 'package:quiver/core.dart';
 
 import '../../../decoder/parser/ranges.dart';
 import '../../../decoder/parser/whitespace.dart';
@@ -110,5 +109,5 @@ class TomlMultilineLiteralString extends TomlMultilineString {
       other is TomlMultilineLiteralString && value == other.value;
 
   @override
-  int get hashCode => hash3(type, stringType, value);
+  int get hashCode => Object.hash(type, stringType, value);
 }