+++ lib/src/util/container/pair.dart
@@ -1,7 +1,6 @@
 library toml.src.util.container.pair;
 
 import 'package:meta/meta.dart';
-import 'package:quiver/core.dart';
 
 /// Utility class that stores a pair of values of types `F` and `S`.
 @immutable
@@ -16,7 +15,7 @@ class Pair<F, S> {
   Pair(this.first, this.second);
 
   @override
-  int get hashCode => hash2(first, second);
+  int get hashCode => Object.hash(first, second);
 
   @override
   bool operator ==(Object other) =>